backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
PyMySQL==1.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openpyxl
pandas
